requirements.txt
amqp==5.1.1
aniso8601==9.0.1
asgiref==3.5.2
async-timeout==4.0.2
attrs==22.2.0
billiard==3.6.4.0
celery==5.2.7
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.6
coreapi==2.3.3
coreschema==0.0.4
coverage==7.2.2
cryptography==39.0.1
defusedxml==0.7.1
Django==4.1.2
django-crispy-forms==1.14.0
django-filter==22.1
django-templated-mail==1.1.1
django_debug_toolbar==3.8.1
djangorestframework==3.14.0
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
exceptiongroup==1.1.1
flake8==6.0.0
flower==1.2.0
graphene==3.2.2
graphene-django==3.1.1
graphql-core==3.2.3
graphql-relay==3.2.0
humanize==4.6.0
idna==3.4
iniconfig==2.0.0
isort==5.11.4
itypes==1.2.0
Jinja2==3.1.2
kombu==5.2.4
Markdown==3.4.1
MarkupSafe==2.1.2
mccabe==0.7.0
mypy==0.991
mypy-extensions==0.4.3
oauthlib==3.2.2
packaging==23.0
Pillow==9.2.0
pluggy==1.0.0
prometheus-client==0.16.0
promise==2.3
prompt-toolkit==3.0.37
psycopg2==2.9.5
pycodestyle==2.10.0
pycparser==2.21
pyflakes==3.0.1
PyJWT==2.6.0
python3-openid==3.2.0
pytz==2022.7.1
redis==4.5.1
requests==2.28.2
requests-oauthlib==1.3.1
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.3.0
sqlparse==0.4.3
text-unidecode==1.3
tomli==2.0.1
tornado==6.2
typing_extensions==4.4.0
tzdata==2022.5
uritemplate==4.1.1
urllib3==1.26.14
vine==5.0.0
wcwidth==0.2.6
